requirements.txt
torch>=2.0
torchvision>=0.15
scikit-learn
matplotlib
Pillow
flask
gunicorn
